+++ CHANGES.rst
@@ -18,6 +18,9 @@ New features:
 - Download XLSX version of saved data.
   [mathias.leimgruber] (#285)
 
+- Add option to append formdata as xlsx as well.
+  [mathias.leimgruber] (#287)
+
 
 Bug fixes: